refactor: schedule폴더 이름 home으로 변경
