+++ src/screens/employer/EmployerHomeScreen.tsx
@@ -18,9 +18,9 @@ import WeeklyDateBar from "../../components/common/WeeklyDateBar";
 import BottomSheetModal from "../../components/common/BottomSheetModal";
 import MonthlyCalendar from "../../components/common/MonthlyCalendar";
 import { colors } from "../../constants/colors";
-import EmployerTimeline from "../../components/employer/schedule/EmployerTimeline";
-import EmployerWorkerListSection from "../../components/employer/schedule/EmployerWorkerListSection";
-import EmployerAddWorkModal from "../../components/employer/schedule/EmployerAddWorkModal";
+import EmployerTimeline from "../../components/employer/home/EmployerTimeline";
+import EmployerWorkerListSection from "../../components/employer/home/EmployerWorkerListSection";
+import EmployerAddWorkModal from "../../components/employer/home/EmployerAddWorkModal";
 import { useWorkplaceManagement } from "../../hooks/employer/useWorkplaceManagement";
 import useEmployerDailyWorkRecords from "../../hooks/employer/useEmployerDailyWorkRecords";
 import type { WorkplaceDetails } from "../../api/employer/types";
